Create aws-utils package
