+++ packages/aws-utils/README.md
@@ -1,0 +1,18 @@
+# Salto - S3 Utils
+
+S3 utils for salto.io.
+
+
+### Build instructions
+
+```
+yarn
+yarn build
+```
+
+### Usage instructions
+
+Add aws-utils to package.json: `yarn add aws-utils`
+Run `yarn install`
+
+import the needed modules from 'aws-utils'
