+++ packages/core/package.json
@@ -32,8 +32,8 @@
     "generate": "./generate.sh"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.171.0",
-    "@aws-sdk/middleware-retry": "^3.171.0",
+    "@aws-sdk/client-s3": "3.198.0",
+    "@salto-io/aws-utils": "0.3.25",
     "@salto-io/adapter-api": "0.3.25",
     "@salto-io/adapter-components": "0.3.25",
     "@salto-io/adapter-utils": "0.3.25",
