+++ packages/core/tsconfig.json
@@ -15,6 +15,7 @@
         { "path": "../adapter-api" },
         { "path": "../adapter-utils" },
         { "path": "../adapter-components" },
+        { "path": "../aws-utils" },
         { "path": "../dag" },
         { "path": "../dummy-adapter" },
         { "path": "../file" },
