+++ packages/persistent-pool/package.json
@@ -27,8 +27,8 @@
     "lint-fix": "yarn lint --fix"
   },
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.171.0",
-    "@aws-sdk/lib-dynamodb": "^3.171.0",
+    "@aws-sdk/client-dynamodb": "3.198.0",
+    "@aws-sdk/lib-dynamodb": "3.198.0",
     "@salto-io/logging": "0.3.25",
     "@salto-io/lowerdash": "0.3.25",
     "@salto-io/test-utils": "0.3.25",
@@ -36,7 +36,7 @@
     "uuid": "^8.3.0"
   },
   "devDependencies": {
-    "@aws-sdk/types": "^3.171.0",
+    "@aws-sdk/types": "3.198.0",
     "@types/jest": "^27.4.0",
     "@types/lodash": "^4.14.168",
     "@types/node": "^12.7.1",
